+++ odoo/songs/install/mts/sales.py
@@ -2,21 +2,21 @@
 # Copyright 2017 Camptocamp SA
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html)
 
-from pkg_resources import resource_stream
-
 import anthem
 
-from ...common import req, load_csv
+from ...common import load_csv
+from ...common import set_sale_conditions
 
 
 @anthem.log
-def set_sale_conditions(ctx):
-    """ Set sale description """
+def set_sale_conditions_MTS(ctx):
     company = ctx.env.ref('base.main_company')
-    note = resource_stream(
-        req,
-        'data/install/mts/mts_sale_conditions_2017_04.html')
-    company.sale_note = note.read()
+    conditions = [
+        # lang, filepath
+        ('en_US', 'data/install/mts/mts_sale_conditions_en.html'),
+        ('fr_FR', 'data/install/mts/mts_sale_conditions_fr.html'),
+    ]
+    set_sale_conditions(ctx, company, conditions)
 
 
 @anthem.log
@@ -41,7 +41,7 @@ def import_product_sellers_MTS(ctx):
 @anthem.log
 def main(ctx):
     """ Configuring sales MTS main """
-    set_sale_conditions(ctx)
+    set_sale_conditions_MTS(ctx)
     import_product_category_MTS(ctx)
     import_product_template_MTS(ctx)